--- ssh/sshshell_cmd.go
@@ -1,1 +1,0 @@
-package ssh
